backend/requirements.txt
torch>=2.0
torchvision>=0.15
Pillow
numpy
opencv-python
tqdm
matplotlib
fastapi
uvicorn
python-multipart